+++ lib/storage/data/data.dart
@@ -55,13 +55,14 @@ class Data {
       loanPayments, // 7
       onlineAccounts, // 8
       payees, // 9
-      rentBuildings, // 10
-      rentUnits, // 11
       securities, // 12
       splits, // 13
       stockSplits, // 14
       transactionExtras, // 15
       transactions, // 16
+      // must come after Transactions
+      rentBuildings, // 10
+      rentUnits, // 11
     ];
   } // private constructor